[CHORE] 토큰 로직 검사 필요없는 endpoint PresentController로 수정
